Rename LUKS_Data to generic Encryption_Data with backward compat
The LUKS-specific static data singleton is renamed to use generic
"encryption" naming (Encryption_Data class, encryption_data singleton)
to support future encrypted format types. The old luks_data module is
preserved as a backward-compatibility shim that re-exports from the
new encryption_data module, so existing consumers (anaconda, etc.)
continue to work without changes.
