+++ blivet/static_data/__init__.py
@@ -1,4 +1,5 @@
 from .lvm_info import lvs_info, pvs_info, vgs_info
+from .encryption_data import encryption_data
 from .luks_data import luks_data
 from .mpath_info import mpath_members
 from .stratis_info import stratis_info
